Uploading files to Github
This is just a draft. 
Things done: front end is 80% done. However I might change how things look and the UI if I got some time. The backend has been started but still needs a lot of work. most functionalities has not been implemented. 
Things to be done: finish the frontend and the backend.